fix: macos build geo_family.cc (#5640)

+++ src/server/geo_family.cc
@@ -405,7 +405,7 @@ void SortIfNeeded(GeoArray* ga, Sorting sorting, uint64_t count) {
   };
 
   if (count > 0) {
-    count = std::min(count, ga->size());
+    count = std::min(count, static_cast<uint64_t>(ga->size()));
     std::partial_sort(ga->begin(), ga->begin() + count, ga->end(), comparator);
     ga->resize(count);
   } else {
